websocket/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2=]
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.2.9
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
zope.interface==5.4.0





